pyproject.toml
[project]
name = "domekit"
version = "0.1.0"
description = "Local-first AI runtime with enforced privacy boundaries"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110,<1",
    "uvicorn>=0.29,<1",
    "pydantic>=2.0,<3",
    "httpx>=0.27,<1",
    "PyYAML>=6.0,<7",
    "jsonschema>=4.21,<5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["runtime*", "contracts*", "apps*", "cli*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
